+++ HISTORY.md
@@ -4,6 +4,7 @@ Work in progress
 - Fix position mapping for a match at the first character in a file.
 - Make `yalafi.shell` subprocess cancellable.
 - Fix error due to faulty diagnostic entries from LaTeX Workshop extension
+- Use `type.typeName` to infer diagnostic severity.
 
 
 Version 0.0.1 (2023/03/08)